fix(ci,quality): repair staging tests and clear all code-scanning alerts
CI test fixes:
- test_mri_nifti_io: write .nii.gz via in-memory deflate (MAX_WBITS+16)
  instead of gz* helpers, which Qt's bundled zlib omits on Windows CI.
- testframes: link test executables with --coverage under WITH_CODE_COV
  so the gcov runtime is pulled in for coverage-instrumented plugin libs.
- test_writetofile_status: copy the runtime-loaded scan_writetofile plugin
  next to the test binary and add @executable_path/$ORIGIN to its rpath so
  the loader resolves it regardless of launch method.

Code scanning (all 56 open alerts):
- 40x integer-multiplication-cast-to-long: widen one operand before the
  multiply (Eigen::Index / qsizetype / std::size_t / double / qreal).
- 12x comparison-with-wider-type: widen narrow loop counters to match the
  wider size()/limit type.
- 4x wrong-type-format-argument: cast size() to int for %d specifiers.

+++ src/examples/ex_evoked_grad_amp/main.cpp
@@ -191,13 +191,13 @@ int main(int argc, char *argv[])
     printf("[done]\n");
 
     //Compose the selection name list
-    for(qint16 i=0; i < npair; ++i) {
+    for(qint32 i=0; i < npair; ++i) {
         ch_sel_names.append(p_FiffEvokedSet.info.ch_names.at(i));
     }
 
     //Omit MEG channels but include others
     qint16 k = npair;
-    for(qint16 p=0; p < p_FiffEvokedSet.info.nchan; ++p) {
+    for(qint32 p=0; p < p_FiffEvokedSet.info.nchan; ++p) {
         if((p_FiffEvokedSet.info.channel_type(p) == "grad") || (p_FiffEvokedSet.info.channel_type(p) == "mag")) {
             ++k;
             ch_sel_names.append(p_FiffEvokedSet.info.ch_names.at(p));
